Move and rename interface
